requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.1=cpu_0
absl-py=0.9.0=pypi_0
argparse=1.4.0=pypi_0
asn1crypto=1.2.0=py36_0
astor=0.8.1=pypi_0
astroid=2.3.3=pypi_0
atari-py=0.2.6=pypi_0
attrs=19.3.0=py_0
beautifulsoup4=4.8.2=py36_0
blas=1.0=mkl
bottleneck=1.3.1=py36hdd07704_0
box2d-py=2.3.8=pypi_0
ca-certificates=2019.11.27=0
cachetools=4.0.0=pypi_0
cairo=1.14.12=h8948797_3
cairocffi=1.1.0=pypi_0
certifi=2019.11.28=py36_0
cffi=1.13.2=py36h2e261b9_0
chardet=3.0.4=py36_1003
click=7.0=pypi_0
cloudpickle=1.2.2=pypi_0
colour=0.1.5=pypi_0
cryptography=2.8=py36h1ba5d50_0
cudatoolkit=10.2.89=hfd86e86_0
cycler=0.10.0=pypi_0
cymem=2.0.2=py36hfd86e86_0
cython=0.29.14=pypi_0
cython-blis=0.2.4=py36h516909a_1
dataclasses=0.6=py36_0
dbus=1.13.12=h746ee38_0
decorator=4.4.1=pypi_0
docopt=0.6.2=pypi_0
et-xmlfile=1.0.1=pypi_0
expat=2.2.6=he6710b0_0
fastai=1.0.60=1
fastprogress=0.2.2=py_0
flask=1.1.1=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
future=0.18.2=pypi_0
gast=0.2.2=pypi_0
glfw=1.9.0=pypi_0
glib=2.63.1=h5a9c865_0
google-auth=1.10.0=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-pasta=0.1.8=pypi_0
grpcio=1.26.0=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
gym=0.15.4=pypi_0
h5py=2.10.0=pypi_0
icu=58.2=h9c2bf20_1
idna=2.8=pypi_0
imageio=2.6.1=pypi_0
importlib_metadata=1.3.0=py36_0
intel-openmp=2019.4=243
isort=4.3.21=pypi_0
itsdangerous=1.1.0=pypi_0
jdcal=1.4.1=pypi_0
jinja2=2.10.3=pypi_0
joblib=0.14.1=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py36_0
keras=2.3.1=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.0=pypi_0
keras-rl=0.4.2=pypi_0
kiwisolver=1.1.0=py36he6710b0_0
lazy-object-proxy=1.4.3=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
lockfile=0.12.2=pypi_0
markdown=3.1.1=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.1.2=pypi_0
mccabe=0.6.1=pypi_0
mkl=2019.4=243
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.15=py36ha843d7b_0
mkl_random=1.1.0=py36hd6b4f25_0
more-itertools=8.0.2=py_0
murmurhash=1.0.2=py36he6710b0_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py36hfd86e86_0
numexpr=2.7.0=py36h9e4a6bb_0
numpy=1.18.1=pypi_0
numpy-base=1.17.4=py36hde5b4d6_0
nvidia-ml-py3=7.352.0=pypi_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py36_0
opencv-python=3.4.2.17=pypi_0
openpyxl=3.0.2=pypi_0
openssl=1.1.1d=h7b6447c_3
opt-einsum=3.1.0=pypi_0
packaging=20.0=py_0
pandas=0.25.3=py36he6710b0_0
pcre=8.43=he6710b0_0
pillow=5.2.0=pypi_0
pip=19.3.1=py36_0
pixman=0.38.0=h516909a_1003
plac=0.9.6=py36_0
preshed=2.0.1=py36he6710b0_0
progressbar=2.5=pypi_0
protobuf=3.11.2=pypi_0
psutil=5.6.7=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.7=pypi_0
pycairo=1.17.1=pypi_0
pycparser=2.19=pypi_0
pydub=0.23.0=pypi_0
pyglet=1.3.2=pypi_0
pylint=2.4.4=pypi_0
pyopenssl=19.1.0=py36_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py36h05f1152_2
pyrsistent=0.15.6=py36h7b6447c_0
pysocks=1.7.1=py36_0
python=3.6.10=h0371630_0
python-dateutil=2.8.1=py_0
pytorch=1.3.1=cpu_py36h62f834f_0
pytz=2019.3=py_0
pyyaml=5.2=pypi_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
requests=2.22.0=py36_1
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=pypi_0
scikit-learn=0.22.1=pypi_0
scipy=1.1.0=pypi_0
setuptools=44.0.0=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.13.0=py36_0
sklearn=0.0=pypi_0
soupsieve=1.9.5=py36_0
spacy=2.1.8=py36hc9558a2_0
sqlite=3.30.1=h7b6447c_0
srsly=0.1.0=py36he1b5a44_0
tensorboard=2.0.2=pypi_0
tensorflow=2.0.0=pypi_0
tensorflow-estimator=2.0.1=pypi_0
termcolor=1.1.0=pypi_0
thinc=7.0.8=py36hc9558a2_0
tk=8.6.8=hbc83047_0
torchvision=0.4.2=cpu_py36h9ec355b_0
tornado=6.0.3=py36h7b6447c_0
tqdm=4.24.0=pypi_0
typed-ast=1.4.0=pypi_0
urllib3=1.25.7=py36_0
utils=1.0.0=pypi_0
wasabi=0.2.2=py_0
werkzeug=0.16.0=pypi_0
wheel=0.33.6=py36_0
wrapt=1.11.2=pypi_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zipp=0.6.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
